Create problem 19.py
Its the dynamic approach. Here, you divide your problem into subproblems and solve them and store the results so that you do not have to calculate same thing for a new subproblems. It reduces the time consumption. Although, it may consume memory but here the we are overwriting in the input variable so, no memory consumption at all.
Same approach in problem 67